+++ src/lib/components/modal/contract/EditContractDetails.tsx
@@ -15,5 +15,6 @@ export const EditContractDetails = ({
     subtitle="Filled information below will be saved on Celatone only and able to edit later."
     contractLocalInfo={contractLocalInfo}
     triggerElement={triggerElement}
+    isSave={false}
   />
 );